+++ src/TestSuite/ValidationTestCase.php
@@ -65,8 +65,8 @@ abstract class ValidationTestCase extends TestCase
      * It first verifies that the entity, compiled with passed data, has no errors.
      *
      * Then it removes one field at a time, verifying that the "this field is required" error is generated.
-     * @param array<string, mixed> $data Valid data
-     * @param array<string> $exclude Key to be excluded
+     * @param array $data Valid data
+     * @param string[] $exclude Key to be excluded
      * @return void
      */
     public function assertAllDataAreRequired(array $data, array $exclude = [])