+++ src/Utility/Sitemap/Sitemap.php
@@ -165,7 +165,6 @@ class Sitemap extends SitemapBase
     /**
      * Returns static pages urls
      * @return array<array{loc: string, lastmod: string, priority: string}>
-     * @throws \ErrorException
      */
     public static function staticPages(): array
     {