+++ templates/layout/rss/default.php
@@ -11,18 +11,10 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ *
+ * @var \Feed\View\RssView $this
  */
-if (!isset($documentData)) {
-    $documentData = [];
-}
-
-if (!isset($channelData)) {
-    $channelData = [];
-}
-
-if (!isset($channelData['title'])) {
-    $channelData['title'] = $this->fetch('title');
-}
+$channelData = ($channelData ?? []) + ['title' => $this->fetch('title')];
 
 $channel = $this->Rss->channel([], $channelData, $this->fetch('content'));
-echo $this->Rss->document($documentData, $channel);
+echo $this->Rss->document($documentData ?? [], $channel);